requirements.txt
mysql-connector==2.2.9
pygithub==1.54.1
matplotlib==3.4.1